creation and definition of the IdCadastro entity

+++ src/main/java/com/example/fingeraccess/repository/LeitorBiometricoRepository.java
@@ -4,6 +4,6 @@ import com.example.fingeraccess.entidade.LeitorBiometrico;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeitorBiometricoRepository extends JpaRepository<LeitorBiometrico, Integer> {
+public interface LeitorBiometricoRepository extends JpaRepository<LeitorBiometrico, Long> {
 
 }
